requirements.txt
blinker==1.7.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyPDF2==3.0.1
pyzbar==0.1.9
reportlab==4.1.0
waitress==3.0.0
Werkzeug==3.0.1
